Remove os, to proceed installation and fail at using
Why?
Currently in sales there is a security issue with tar ball being downloaded from S3. As  the tar ball is not available in S3, a hacker could claim and place any arbitrary code.

To fix, added "resolutions" for the dependencies. Once fsevents is added in resolutions, it is not behaving as optional package and failing to install in Linux

As all doors are closed, decided to monkey patch this by enabling for all OS.